headers updated and notification center moved
