+++ packages/optimizely-sdk/lib/utils/event_tag_utils/index.tests.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2017, 2020-2022, Optimizely
+ * Copyright 2017, 2020, 2022, Optimizely
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
