+++ packages/optimizely-sdk/lib/utils/json_schema_validator/index.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright 2016-2017, 2020-2022 Optimizely
+ * Copyright 2016-2017, 2020, 2022 Optimizely
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
