+++ packages/optimizely-sdk/lib/utils/user_profile_service_validator/index.tests.js
@@ -1,5 +1,5 @@
 /****************************************************************************
- * Copyright 2017, 2020-2022, Optimizely, Inc. and contributors                  *
+ * Copyright 2017, 2020, 2022, Optimizely, Inc. and contributors                  *
  *                                                                          *
  * Licensed under the Apache License, Version 2.0 (the "License");          *
  * you may not use this file except in compliance with the License.         *
